Refactor proxy_logging and namespace handlers
