+++ handlers/proxy_logging_test.go
@@ -1,42 +1,47 @@
-package handlers
+package handlers_test
 
 import (
 	"fmt"
 	"io"
-	"net/http"
 	"net/http/httptest"
 	"testing"
 
 	"github.com/gorilla/mux"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/client-go/tools/clientcmd/api"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/kiali/kiali/business"
 	"github.com/kiali/kiali/config"
+	"github.com/kiali/kiali/grafana"
+	"github.com/kiali/kiali/handlers"
+	"github.com/kiali/kiali/istio"
+	"github.com/kiali/kiali/kubernetes/cache"
 	"github.com/kiali/kiali/kubernetes/kubetest"
+	"github.com/kiali/kiali/prometheus/prometheustest"
+	"github.com/kiali/kiali/tracing"
 )
 
 func setupTestLoggingServer(t *testing.T, namespace, pod string) *httptest.Server {
+	conf := config.NewConfig()
+	k8s := kubetest.NewFakeK8sClient(
+		&corev1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: namespace}},
+		&corev1.Pod{ObjectMeta: metav1.ObjectMeta{Name: pod, Namespace: namespace}, Status: corev1.PodStatus{Phase: corev1.PodRunning}},
+	)
+	prom := new(prometheustest.PromClientMock)
+	cf := kubetest.NewFakeClientFactoryWithClient(conf, k8s)
+	cache := cache.NewTestingCacheWithFactory(t, cf, *conf)
+	discovery := istio.NewDiscovery(cf.Clients, cache, conf)
+	cpm := &business.FakeControlPlaneMonitor{}
+	traceLoader := func() tracing.ClientInterface { return nil }
+	grafana := grafana.NewService(conf, cf.GetSAHomeClusterClient())
+
+	handler := handlers.WithFakeAuthInfo(conf, handlers.LoggingUpdate(conf, cache, cf, cpm, prom, traceLoader, grafana, discovery))
 	mr := mux.NewRouter()
-	path := "/api/namespaces/{namespace}/pods/{pod}/logging"
-	authInfo := map[string]*api.AuthInfo{config.Get().KubernetesConfig.ClusterName: {Token: "test"}}
-	mr.HandleFunc(path, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		WithAuthInfo(authInfo, LoggingUpdate)(w, r)
-	}))
+	mr.HandleFunc("/api/namespaces/{namespace}/pods/{pod}/logging", handler)
 
 	ts := httptest.NewServer(mr)
 	t.Cleanup(ts.Close)
-
-	k8s := new(kubetest.K8SClientMock)
-	k8s.On("IsOpenShift").Return(false)
-	k8s.On("IsGatewayAPI").Return(false)
-	k8s.On("SetProxyLogLevel").Return(nil)
-	var fakePod *corev1.Pod
-	k8s.On("GetPod", namespace, pod).Return(fakePod, nil)
-
-	mockClientFactory := kubetest.NewK8SClientFactoryMock(k8s)
-	business.SetWithBackends(mockClientFactory, nil)
 
 	return ts
 }
